Make RichSelectItem generic again
